+++ src/core/mod.rs
@@ -6,7 +6,6 @@ mod err;
 mod message_delegate;
 mod msg;
 mod play_msg;
-mod song_order;
 mod task_msg;
 mod uamp_app;
 
@@ -22,6 +21,5 @@ pub mod query;
 
 pub use self::{
     any_control_msg::*, control_msg::*, data_control_msg::*, err::*,
-    message_delegate::*, msg::*, play_msg::*, song_order::*, task_msg::*,
-    uamp_app::*,
+    message_delegate::*, msg::*, play_msg::*, task_msg::*, uamp_app::*,
 };