+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## future
 - Add new functionality to filters (and, or, brackets).
 - Add shorter aliases for sorting and filtering.
+- Add option to sort songs as part of query
 
 ## v0.5.4
 - Add option to enable/disable color in terminal.